edge/sync-service/requirements.txt
psycopg2-binary
firebase-admin
